Implement_Guardrails_Using_OpenAI_Agent_SDK/pyproject.toml
[project]
name = "implement-guardrails-using-openai-agent-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
